# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BugCategory = (typeof BugCategory)[keyof typeof BugCategory];

export const BugCategory = {
  UI: "UI",
  UX: "UX",
  "Client-side": "Client-side",
  "Server-side": "Server-side",
  "SSR-Hydration": "SSR-Hydration",
  Performance: "Performance",
  Security: "Security",
  Accessibility: "Accessibility",
  Product_Content: "Product Content",
  Other: "Other",
} as const;

export type BugSeverity = (typeof BugSeverity)[keyof typeof BugSeverity];

export const BugSeverity = {
  Critical: "Critical",
  High: "High",
  Medium: "Medium",
  Low: "Low",
  Cosmetic: "Cosmetic",
} as const;

export type BugStatus = (typeof BugStatus)[keyof typeof BugStatus];

export const BugStatus = {
  Draft: "Draft",
  Submitted: "Submitted",
  Confirmed: "Confirmed",
  Duplicate: "Duplicate",
  Invalid: "Invalid",
} as const;

export interface Bug {
  id: number;
  title: string;
  category: BugCategory;
  severity: BugSeverity;
  status: BugStatus;
  environment: string;
  stepsToReproduce: string;
  expectedBehaviour: string;
  actualBehaviour: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm: boolean;
  formSubmissionUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateBugBodyCategory =
  (typeof CreateBugBodyCategory)[keyof typeof CreateBugBodyCategory];

export const CreateBugBodyCategory = {
  UI: "UI",
  UX: "UX",
  "Client-side": "Client-side",
  "Server-side": "Server-side",
  "SSR-Hydration": "SSR-Hydration",
  Performance: "Performance",
  Security: "Security",
  Accessibility: "Accessibility",
  Product_Content: "Product Content",
  Other: "Other",
} as const;

export type CreateBugBodySeverity =
  (typeof CreateBugBodySeverity)[keyof typeof CreateBugBodySeverity];

export const CreateBugBodySeverity = {
  Critical: "Critical",
  High: "High",
  Medium: "Medium",
  Low: "Low",
  Cosmetic: "Cosmetic",
} as const;

export type CreateBugBodyStatus =
  (typeof CreateBugBodyStatus)[keyof typeof CreateBugBodyStatus];

export const CreateBugBodyStatus = {
  Draft: "Draft",
  Submitted: "Submitted",
  Confirmed: "Confirmed",
  Duplicate: "Duplicate",
  Invalid: "Invalid",
} as const;

export interface CreateBugBody {
  title: string;
  category: CreateBugBodyCategory;
  severity: CreateBugBodySeverity;
  status?: CreateBugBodyStatus;
  environment: string;
  stepsToReproduce: string;
  expectedBehaviour: string;
  actualBehaviour: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm?: boolean;
  formSubmissionUrl?: string | null;
}

export type UpdateBugBodyCategory =
  (typeof UpdateBugBodyCategory)[keyof typeof UpdateBugBodyCategory];

export const UpdateBugBodyCategory = {
  UI: "UI",
  UX: "UX",
  "Client-side": "Client-side",
  "Server-side": "Server-side",
  "SSR-Hydration": "SSR-Hydration",
  Performance: "Performance",
  Security: "Security",
  Accessibility: "Accessibility",
  Product_Content: "Product Content",
  Other: "Other",
} as const;

export type UpdateBugBodySeverity =
  (typeof UpdateBugBodySeverity)[keyof typeof UpdateBugBodySeverity];

export const UpdateBugBodySeverity = {
  Critical: "Critical",
  High: "High",
  Medium: "Medium",
  Low: "Low",
  Cosmetic: "Cosmetic",
} as const;

export type UpdateBugBodyStatus =
  (typeof UpdateBugBodyStatus)[keyof typeof UpdateBugBodyStatus];

export const UpdateBugBodyStatus = {
  Draft: "Draft",
  Submitted: "Submitted",
  Confirmed: "Confirmed",
  Duplicate: "Duplicate",
  Invalid: "Invalid",
} as const;

export interface UpdateBugBody {
  title?: string;
  category?: UpdateBugBodyCategory;
  severity?: UpdateBugBodySeverity;
  status?: UpdateBugBodyStatus;
  environment?: string;
  stepsToReproduce?: string;
  expectedBehaviour?: string;
  actualBehaviour?: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm?: boolean;
  formSubmissionUrl?: string | null;
}

export type BugStatsByCategoryItem = {
  category: string;
  count: number;
};

export type BugStatsBySeverityItem = {
  severity: string;
  count: number;
};

export type BugStatsByStatusItem = {
  status: string;
  count: number;
};

export interface BugStats {
  total: number;
  byCategory: BugStatsByCategoryItem[];
  bySeverity: BugStatsBySeverityItem[];
  byStatus: BugStatsByStatusItem[];
  rewardEligible: number;
}

export type ListBugsParams = {
  category?: string;
  severity?: string;
  search?: string;
  status?: string;
};

export type GetRecentBugsParams = {
  limit?: number;
};