# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns a list of all bug reports
 * @summary List all bugs
 */
export const ListBugsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  severity: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const ListBugsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.enum([
    "UI",
    "UX",
    "Client-side",
    "Server-side",
    "SSR-Hydration",
    "Performance",
    "Security",
    "Accessibility",
    "Product Content",
    "Other",
  ]),
  severity: zod.enum(["Critical", "High", "Medium", "Low", "Cosmetic"]),
  status: zod.enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"]),
  environment: zod.string(),
  stepsToReproduce: zod.string(),
  expectedBehaviour: zod.string(),
  actualBehaviour: zod.string(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean(),
  formSubmissionUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBugsResponse = zod.array(ListBugsResponseItem);

/**
 * @summary Create a bug report
 */
export const CreateBugBody = zod.object({
  title: zod.string(),
  category: zod.enum([
    "UI",
    "UX",
    "Client-side",
    "Server-side",
    "SSR-Hydration",
    "Performance",
    "Security",
    "Accessibility",
    "Product Content",
    "Other",
  ]),
  severity: zod.enum(["Critical", "High", "Medium", "Low", "Cosmetic"]),
  status: zod
    .enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"])
    .optional(),
  environment: zod.string(),
  stepsToReproduce: zod.string(),
  expectedBehaviour: zod.string(),
  actualBehaviour: zod.string(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean().optional(),
  formSubmissionUrl: zod.string().nullish(),
});

/**
 * @summary Get a single bug
 */
export const GetBugParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBugResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.enum([
    "UI",
    "UX",
    "Client-side",
    "Server-side",
    "SSR-Hydration",
    "Performance",
    "Security",
    "Accessibility",
    "Product Content",
    "Other",
  ]),
  severity: zod.enum(["Critical", "High", "Medium", "Low", "Cosmetic"]),
  status: zod.enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"]),
  environment: zod.string(),
  stepsToReproduce: zod.string(),
  expectedBehaviour: zod.string(),
  actualBehaviour: zod.string(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean(),
  formSubmissionUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a bug report
 */
export const UpdateBugParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBugBody = zod.object({
  title: zod.string().optional(),
  category: zod
    .enum([
      "UI",
      "UX",
      "Client-side",
      "Server-side",
      "SSR-Hydration",
      "Performance",
      "Security",
      "Accessibility",
      "Product Content",
      "Other",
    ])
    .optional(),
  severity: zod
    .enum(["Critical", "High", "Medium", "Low", "Cosmetic"])
    .optional(),
  status: zod
    .enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"])
    .optional(),
  environment: zod.string().optional(),
  stepsToReproduce: zod.string().optional(),
  expectedBehaviour: zod.string().optional(),
  actualBehaviour: zod.string().optional(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean().optional(),
  formSubmissionUrl: zod.string().nullish(),
});

export const UpdateBugResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.enum([
    "UI",
    "UX",
    "Client-side",
    "Server-side",
    "SSR-Hydration",
    "Performance",
    "Security",
    "Accessibility",
    "Product Content",
    "Other",
  ]),
  severity: zod.enum(["Critical", "High", "Medium", "Low", "Cosmetic"]),
  status: zod.enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"]),
  environment: zod.string(),
  stepsToReproduce: zod.string(),
  expectedBehaviour: zod.string(),
  actualBehaviour: zod.string(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean(),
  formSubmissionUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a bug report
 */
export const DeleteBugParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * Returns aggregated bug stats by category, severity, and status
 * @summary Get bug statistics summary
 */
export const GetBugStatsResponse = zod.object({
  total: zod.number(),
  byCategory: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
  bySeverity: zod.array(
    zod.object({
      severity: zod.string(),
      count: zod.number(),
    }),
  ),
  byStatus: zod.array(
    zod.object({
      status: zod.string(),
      count: zod.number(),
    }),
  ),
  rewardEligible: zod.number(),
});

/**
 * @summary Get recently added bugs
 */
export const getRecentBugsQueryLimitDefault = 5;

export const GetRecentBugsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentBugsQueryLimitDefault),
});

export const GetRecentBugsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  category: zod.enum([
    "UI",
    "UX",
    "Client-side",
    "Server-side",
    "SSR-Hydration",
    "Performance",
    "Security",
    "Accessibility",
    "Product Content",
    "Other",
  ]),
  severity: zod.enum(["Critical", "High", "Medium", "Low", "Cosmetic"]),
  status: zod.enum(["Draft", "Submitted", "Confirmed", "Duplicate", "Invalid"]),
  environment: zod.string(),
  stepsToReproduce: zod.string(),
  expectedBehaviour: zod.string(),
  actualBehaviour: zod.string(),
  rootCause: zod.string().nullish(),
  suggestedFix: zod.string().nullish(),
  submittedToForm: zod.boolean(),
  formSubmissionUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentBugsResponse = zod.array(GetRecentBugsResponseItem);