# lib/api-zod/src/generated/types/bug.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { BugCategory } from "./bugCategory";
import type { BugSeverity } from "./bugSeverity";
import type { BugStatus } from "./bugStatus";

export interface Bug {
  id: number;
  title: string;
  category: BugCategory;
  severity: BugSeverity;
  status: BugStatus;
  environment: string;
  stepsToReproduce: string;
  expectedBehaviour: string;
  actualBehaviour: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm: boolean;
  formSubmissionUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}