# lib/api-zod/src/generated/types/bugStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { BugStatsByCategoryItem } from "./bugStatsByCategoryItem";
import type { BugStatsBySeverityItem } from "./bugStatsBySeverityItem";
import type { BugStatsByStatusItem } from "./bugStatsByStatusItem";

export interface BugStats {
  total: number;
  byCategory: BugStatsByCategoryItem[];
  bySeverity: BugStatsBySeverityItem[];
  byStatus: BugStatsByStatusItem[];
  rewardEligible: number;
}