# lib/api-zod/src/generated/types/createBugBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBugBodyCategory } from "./createBugBodyCategory";
import type { CreateBugBodySeverity } from "./createBugBodySeverity";
import type { CreateBugBodyStatus } from "./createBugBodyStatus";

export interface CreateBugBody {
  title: string;
  category: CreateBugBodyCategory;
  severity: CreateBugBodySeverity;
  status?: CreateBugBodyStatus;
  environment: string;
  stepsToReproduce: string;
  expectedBehaviour: string;
  actualBehaviour: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm?: boolean;
  formSubmissionUrl?: string | null;
}