# lib/api-zod/src/generated/types/updateBugBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Maya Bug Hunt Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBugBodyCategory } from "./updateBugBodyCategory";
import type { UpdateBugBodySeverity } from "./updateBugBodySeverity";
import type { UpdateBugBodyStatus } from "./updateBugBodyStatus";

export interface UpdateBugBody {
  title?: string;
  category?: UpdateBugBodyCategory;
  severity?: UpdateBugBodySeverity;
  status?: UpdateBugBodyStatus;
  environment?: string;
  stepsToReproduce?: string;
  expectedBehaviour?: string;
  actualBehaviour?: string;
  rootCause?: string | null;
  suggestedFix?: string | null;
  submittedToForm?: boolean;
  formSubmissionUrl?: string | null;
}